# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Loadbyton API
  version: 1.0.0
  description: UAE road freight marketplace — jobs, bids, escrow, payouts
servers:
  - url: http://localhost:4000
    description: Local
  - url: https://api.loadbyton.com
    description: Production (proxied from https://loadbyton.com/api/* per vercel.json)
...
paths:
...
  /api/jobs:
    get:
      tags: [jobs]
      summary: List jobs
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: offset, in: query, schema: { type: integer } }
        - { name: mine, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: equipmentType, in: query, schema: { type: string } }
        - { name: escrowStatus, in: query, schema: { type: string } }
        - { name: shipmentType, in: query, schema: { type: string } }
      responses:
        '200': { description: ok }
        '401': { description: unauthenticated }
    post:
      tags: [jobs]
      summary: Create job
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreate' }
      responses:
        '201': { description: created }
        '403': { description: pending approval or role }
  /api/jobs/{id}:
    get:
      tags: [jobs]
      summary: Get job detail with bids and documents
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: ok }
        '404': { description: not found }
    patch:
      tags: [jobs]
      summary: Edit job while OPEN (no pending bids)
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetPriceAed: { type: number }
                notes: { type: string }
                deliveryAddress: { type: string }
      responses:
        '200': { description: updated }
        '403': { description: not owner or not OPEN }
...
  /api/jobs/{id}/documents:
    post:
      tags: [jobs]
      summary: Attach document to job (shipper or awarded carrier only)
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentCreate' }
      responses:
        '201': { description: created }
        '403': { description: not party }
    get:
      tags: [jobs]
      summary: List visible documents (filtered per role)
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: ok }
  /api/jobs/{id}/documents/{docId}/file:
    get:
      tags: [jobs]
      summary: Download document file (stream or redirect)
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: docId, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: file stream }
        '302': { description: redirect to file_url }
        '404': { description: not found }
...
components:
  schemas:
...
    JobCreate:
      type: object
      required: [pickupTerminal, deliveryArea, deliveryAddress, readyAt, deadline]
      properties:
        containerSize: { type: string }
        containerType: { type: string }
        pickupTerminal: { type: string }
        deliveryArea: { type: string }
        deliveryAddress: { type: string }
        readyAt: { type: string, format: date-time }
        deadline: { type: string, format: date-time }
        maxBudgetAed: { type: number }
        targetPriceAed: { type: number }
        shipmentType: { type: string, enum: [IMPORT, EXPORT, LOCAL] }
        equipmentType: { type: string }
        cargoWeightTons: { type: number }
        notes: { type: string }
...
    DocumentCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        docType: { type: string, enum: [CUSTOMS, RECEIPT, POD, LICENCE, INSURANCE, PACKING_LIST, OTHER, DO, BOE, GATE_PASS, POD_TEMPLATE, INSPECTION_PROOF] }
        fileUrl: { type: string, format: uri }
        fileBase64: { type: string, contentEncoding: base64 }
        mimeType: { type: string }
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: lb_session }